requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
faiss-cpu==1.7.4
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.3
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
langchain-openai==0.0.6
langcodes==3.3.0
langsmith==0.0.87
llama-index==0.10.5
llama-index-agent-openai==0.1.1
llama-index-core==0.10.5
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.2
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
portalocker==2.8.2
preshed==3.0.9
protobuf==4.25.3
pydantic==2.6.1
pydantic_core==2.16.2
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.3
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wasabi==1.1.2
weasel==0.3.4
wrapt==1.16.0
yarl==1.9.4
